services/analytics/pyproject.toml
[project]
name = "analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "shubh305", email = "shubham.305singh@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "aiokafka>=0.13.0",
    "clickhouse-connect>=0.10.0",
    "fastapi>=0.128.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "tenacity>=8.2.3",
    "prometheus-client>=0.19.0",
    "kafka-python>=2.0.2",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-kafka-python>=0.43b0",
    "pydantic-settings>=2.12.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
analytics = "analytics:main"

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"
